+++ Scraps.GUI/Forms/WebViewForm.Designer.cs
@@ -1,0 +1,70 @@
+﻿
+namespace Scraps.GUI.Forms
+{
+    partial class WebViewForm
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(WebViewForm));
+            this._WebBrowser = new Microsoft.Web.WebView2.WinForms.WebView2();
+            ((System.ComponentModel.ISupportInitialize)(this._WebBrowser)).BeginInit();
+            this.SuspendLayout();
+            // 
+            // _WebBrowser
+            // 
+            this._WebBrowser.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
+            | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this._WebBrowser.CreationProperties = null;
+            this._WebBrowser.DefaultBackgroundColor = System.Drawing.Color.White;
+            this._WebBrowser.Location = new System.Drawing.Point(0, 0);
+            this._WebBrowser.Name = "_WebBrowser";
+            this._WebBrowser.Size = new System.Drawing.Size(1264, 681);
+            this._WebBrowser.TabIndex = 0;
+            this._WebBrowser.ZoomFactor = 1D;
+            // 
+            // WebViewForm
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ClientSize = new System.Drawing.Size(1264, 681);
+            this.Controls.Add(this._WebBrowser);
+            this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
+            this.MinimizeBox = false;
+            this.Name = "WebViewForm";
+            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+            this.Text = "WebViewForm";
+            ((System.ComponentModel.ISupportInitialize)(this._WebBrowser)).EndInit();
+            this.ResumeLayout(false);
+
+        }
+
+        #endregion
+
+        private Microsoft.Web.WebView2.WinForms.WebView2 _WebBrowser;
+    }
+}